+++ default_handlers/handlers.php
@@ -12,4 +12,6 @@ return [
         ['Content-Type' => 'text/plain'],
         'OK'
     ),
+    ...require(__DIR__.'/amazon_ses_v2.php'),
+    ...require(__DIR__.'/sendgrid_v3_send.php'),
 ];